requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
numpy==1.21.0
pandas==1.3.0
python-dotenv==0.19.0
python-multipart==0.0.5
supabase==0.7.1
openpyxl==3.0.9
gunicorn==20.1.0